Make banner only appear on homepage of entire website
In the previous commit, I made it appear on the homepages of all the
sub-sites. However, even then, it did not appear on the non-home pages
of those sub-sites. This approach requires less code in the GitHub
Actions workflow, and less custom styling (no need to make it work
well with the ReadTheDocs Sphinx theme).