Remove the trailing colons from the navigation section names
I suspect someone added these because the section names used to appear
in the content area of the home page, and the person did not think the
increased size and weight of the headers sufficiently identified them
as headers of the links that followed them.

Since the section names no longer appear in the content area of the home
page, I am removing those colons here.